# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Future Investment Practice Game API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface CreateGameRequest {
  /** Starting cash amount in dollars */
  startingCash: number;
  /** Player name */
  playerName: string;
}

export type TradeRequestAction =
  (typeof TradeRequestAction)[keyof typeof TradeRequestAction];

export const TradeRequestAction = {
  buy: "buy",
  sell: "sell",
} as const;

export interface TradeRequest {
  /** Stock ticker symbol */
  ticker: string;
  action: TradeRequestAction;
  /** Number of shares to buy or sell */
  shares: number;
}

export interface StockInfo {
  ticker: string;
  name: string;
  sector: string;
  description: string;
  /** Stock volatility factor (higher = more volatile) */
  volatility: number;
  /** Annual dividend yield as decimal (e.g. 0.02 = 2%) */
  dividendYield: number;
}

export interface StockHolding {
  ticker: string;
  shares: number;
  avgCostBasis: number;
  currentPrice: number;
  totalValue: number;
  gainLoss: number;
  gainLossPercent: number;
}

export interface StockPrice {
  ticker: string;
  price: number;
  open: number;
  high: number;
  low: number;
  change: number;
  changePercent: number;
  volume: number;
}

export interface PortfolioSnapshot {
  day: number;
  date: string;
  totalValue: number;
  cashBalance: number;
  investedValue: number;
}

export interface TradeRecord {
  day: number;
  date: string;
  ticker: string;
  action: string;
  shares: number;
  price: number;
  total: number;
}

export type GameStateMarketSentiment =
  (typeof GameStateMarketSentiment)[keyof typeof GameStateMarketSentiment];

export const GameStateMarketSentiment = {
  bullish: "bullish",
  bearish: "bearish",
  neutral: "neutral",
} as const;

export interface GameState {
  sessionId: string;
  playerName: string;
  currentDay: number;
  currentDate: string;
  startingCash: number;
  cashBalance: number;
  holdings: StockHolding[];
  stockPrices: StockPrice[];
  portfolioHistory: PortfolioSnapshot[];
  tradeHistory: TradeRecord[];
  totalPortfolioValue: number;
  totalGainLoss: number;
  totalGainLossPercent: number;
  marketSentiment: GameStateMarketSentiment;
  newsEvents: string[];
}