# lib/api-zod/src/generated/types/createGameRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Future Investment Practice Game API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateGameRequest {
  /** Starting cash amount in dollars */
  startingCash: number;
  /** Player name */
  playerName: string;
}