# lib/api-zod/src/generated/types/gameState.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Future Investment Practice Game API
 * OpenAPI spec version: 0.1.0
 */
import type { GameStateMarketSentiment } from "./gameStateMarketSentiment";
import type { PortfolioSnapshot } from "./portfolioSnapshot";
import type { StockHolding } from "./stockHolding";
import type { StockPrice } from "./stockPrice";
import type { TradeRecord } from "./tradeRecord";

export interface GameState {
  sessionId: string;
  playerName: string;
  currentDay: number;
  currentDate: string;
  startingCash: number;
  cashBalance: number;
  holdings: StockHolding[];
  stockPrices: StockPrice[];
  portfolioHistory: PortfolioSnapshot[];
  tradeHistory: TradeRecord[];
  totalPortfolioValue: number;
  totalGainLoss: number;
  totalGainLossPercent: number;
  marketSentiment: GameStateMarketSentiment;
  newsEvents: string[];
}