# lib/api-zod/src/generated/types/portfolioSnapshot.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Future Investment Practice Game API
 * OpenAPI spec version: 0.1.0
 */

export interface PortfolioSnapshot {
  day: number;
  date: string;
  totalValue: number;
  cashBalance: number;
  investedValue: number;
}