# lib/api-zod/src/generated/types/stockInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Future Investment Practice Game API
 * OpenAPI spec version: 0.1.0
 */

export interface StockInfo {
  ticker: string;
  name: string;
  sector: string;
  description: string;
  /** Stock volatility factor (higher = more volatile) */
  volatility: number;
  /** Annual dividend yield as decimal (e.g. 0.02 = 2%) */
  dividendYield: number;
}