# lib/api-zod/src/generated/types/stockPrice.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Future Investment Practice Game API
 * OpenAPI spec version: 0.1.0
 */

export interface StockPrice {
  ticker: string;
  price: number;
  open: number;
  high: number;
  low: number;
  change: number;
  changePercent: number;
  volume: number;
}