# lib/api-zod/src/generated/types/tradeRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Future Investment Practice Game API
 * OpenAPI spec version: 0.1.0
 */

export interface TradeRecord {
  day: number;
  date: string;
  ticker: string;
  action: string;
  shares: number;
  price: number;
  total: number;
}