# lib/api-zod/src/generated/types/tradeRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Future Investment Practice Game API
 * OpenAPI spec version: 0.1.0
 */
import type { TradeRequestAction } from "./tradeRequestAction";

export interface TradeRequest {
  /** Stock ticker symbol */
  ticker: string;
  action: TradeRequestAction;
  /** Number of shares to buy or sell */
  shares: number;
}